AI: shared settings cog menu on both surfaces; Component editor gets it (#40)
The Component editor ("Edit with AI" / AIBuilderPage) had no settings cog — its
conversation export sat in a standalone Download OverflowMenu and it had no
expand-tool-calls control. The Dashboard Assistant has a richer cog menu. Bring
them to parity via a shared component.

- shared/AgentSettingsMenu.{jsx,scss}: the cog popover, generalized from
  AssistantSettingsMenu. Every item is OPTIONAL (renders only when its handler
  is passed), so each surface shows what it can wire: the Component editor gets
  Export Markdown / Export JSON / Expand tool calls by default; the Assistant
  additionally gets Clear chat.
- "Show token usage" REMOVED from the menu on BOTH surfaces — the live
  per-session display was never wired (issue #55), so a non-functional toggle
  shouldn't show. The per-user daily usage buckets already exist server-side
  (chat_usage); #55 is now purely the display.
- AssistantSettingsMenu is a thin wrapper over the shared menu (accepts +
  ignores the old token-usage props so its call site is unchanged); the
  duplicated .assistant-settings-menu SCSS + the token-usage placeholder
  (render + SCSS) were deleted.
- AIBuilderPage: standalone Download menu replaced with AgentSettingsMenu;
  added a session-scoped expandToolCalls pref threaded into the tool-call
  cards' defaultOpen.

Fourth shared agent component (after tool-call row, exporter, welcome) — both
AI surfaces converge. Build 2002.
