+++ client/build.json
@@ -1,3 +1,3 @@
 {
-  "buildNumber": 2001
+  "buildNumber": 2002
 }
