+++ client/src/components/shared/AgentSettingsMenu.scss
@@ -1,0 +1,79 @@
+// Copyright (c) 2026 TRV Enterprises LLC
+// Licensed under Apache 2.0
+// See LICENSE file for details.
+
+// Shared cog settings-menu (AgentSettingsMenu), used by both the Dashboard
+// Assistant and the Component agent. Theme-neutral Carbon tokens only.
+
+@use '@carbon/react/scss/spacing' as spacing;
+
+.agent-settings-menu__wrap {
+  position: relative;
+  display: inline-flex;
+}
+
+.agent-settings-menu {
+  position: absolute;
+  top: 100%;
+  right: 0;
+  margin-top: 4px;
+  min-width: 16rem;
+  background: var(--cds-layer);
+  border: 1px solid var(--cds-border-subtle-01);
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.4);
+  z-index: 9000;
+  padding: spacing.$spacing-02 0;
+  display: flex;
+  flex-direction: column;
+
+  &__item {
+    display: flex;
+    align-items: center;
+    gap: spacing.$spacing-03;
+    width: 100%;
+    white-space: nowrap;
+    background: transparent;
+    border: none;
+    color: var(--cds-text-primary);
+    font: inherit;
+    font-size: 0.875rem;
+    padding: spacing.$spacing-03 spacing.$spacing-05;
+    cursor: pointer;
+    text-align: left;
+
+    &:hover:not(:disabled) {
+      background: var(--cds-layer-hover);
+    }
+
+    &:disabled {
+      color: var(--cds-text-disabled);
+      cursor: not-allowed;
+
+      svg {
+        color: var(--cds-icon-disabled);
+      }
+    }
+
+    svg {
+      flex: 0 0 16px;
+      color: var(--cds-icon-secondary);
+    }
+  }
+
+  // Toggle rows lead with a checkmark column so on/off prefs read at a glance.
+  &__check {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    flex: 0 0 16px;
+    width: 16px;
+    height: 16px;
+    color: var(--cds-icon-primary);
+  }
+
+  &__divider {
+    height: 1px;
+    background: var(--cds-border-subtle-01);
+    margin: spacing.$spacing-02 0;
+  }
+}
